requirements.txt
# RDIC Research Project Dependencies
# Day 1 - Environment Setup

# Core LLM Libraries
llama-cpp-python>=0.2.0  # Local Llama inference with GGUF support
anthropic>=0.25.0  # Claude API
openai>=1.0.0  # DeepSeek R1 API (OpenAI-compatible)

# Data Processing
numpy>=1.24.0
pandas>=2.0.0
python-dotenv>=1.0.0

# Evaluation & NLP
sentence-transformers>=2.2.0  # For embedding router
scikit-learn>=1.3.0  # For metrics (ARI, NMI)

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Utilities
tqdm>=4.65.0
jupyter>=1.0.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
